Create README.md file
Hey all, this is an ecommerce project built with Html, CSS, JavaScript in the front end with Node JS for its backend programming utilizing Mongo DB to store the data for easy retrieval with queries. The project follows MVC architecture and leveraging the Express framework. For extended functionalities, third party API's like Twilio for OTP functionality, Axios for data fetching and Razorpay for seamless payment integration.